+++ zbus/src/connection/socket/mod.rs
@@ -12,9 +12,20 @@ use std::io;
 use std::sync::Arc;
 use tracing::trace;
 
-use crate::{fdo::ConnectionCredentials, Message};
+use crate::{
+    fdo::ConnectionCredentials,
+    message::{
+        header::{MAX_MESSAGE_SIZE, MIN_MESSAGE_SIZE},
+        PrimaryHeader,
+    },
+    padding_for_8_bytes, Message,
+};
 #[cfg(unix)]
 use std::os::fd::{AsFd, BorrowedFd, OwnedFd};
+use zvariant::{
+    serialized::{self, Context},
+    Endian,
+};
 
 #[cfg(unix)]
 type RecvmsgResult = io::Result<(usize, Vec<OwnedFd>)>;
@@ -50,11 +61,124 @@ pub trait Socket {
 /// See [`Socket`] for more details.
 #[async_trait::async_trait]
 pub trait ReadHalf: std::fmt::Debug + Send + Sync + 'static {
-    /// Attempt to receive a message from the socket.
+    /// Receive a message on the socket.
+    ///
+    /// This is the higher-level method to receive a full D-Bus message.
+    ///
+    /// The default implementation uses `recvmsg` to receive the message. Implementers should
+    /// override either this or `recvmsg`. Note that if you override this method, zbus will not be
+    /// able perform an authentication handshake and hence will skip the handshake. Therefore your
+    /// implementation will only be useful for pre-authenticated connections or connections that do
+    /// not require authentication.
+    ///
+    /// # Parameters
+    ///
+    /// - `seq`: The sequence number of the message. The returned message should have this sequence.
+    /// - `already_received_bytes`: Sometimes, zbus already received some bytes from the socket
+    ///   belonging to the message (as part of the connection handshake process). This is the buffer
+    ///   containing those bytes (if any). If you're implementing this method, most likely you can
+    ///   safely ignore this parameter.
+    async fn receive_message(
+        &mut self,
+        seq: u64,
+        already_received_bytes: Option<Vec<u8>>,
+    ) -> crate::Result<Message> {
+        let mut bytes =
+            already_received_bytes.unwrap_or_else(|| Vec::with_capacity(MIN_MESSAGE_SIZE));
+        let mut pos = bytes.len();
+        #[cfg(unix)]
+        let mut fds = vec![];
+        if pos < MIN_MESSAGE_SIZE {
+            bytes.resize(MIN_MESSAGE_SIZE, 0);
+            // We don't have enough data to make a proper message header yet.
+            // Some partial read may be in raw_in_buffer, so we try to complete it
+            // until we have MIN_MESSAGE_SIZE bytes
+            //
+            // Given that MIN_MESSAGE_SIZE is 16, this codepath is actually extremely unlikely
+            // to be taken more than once
+            while pos < MIN_MESSAGE_SIZE {
+                let res = self.recvmsg(&mut bytes[pos..]).await?;
+                let len = {
+                    #[cfg(unix)]
+                    {
+                        fds.extend(res.1);
+                        res.0
+                    }
+                    #[cfg(not(unix))]
+                    {
+                        res
+                    }
+                };
+                pos += len;
+                if len == 0 {
+                    return Err(std::io::Error::new(
+                        std::io::ErrorKind::UnexpectedEof,
+                        "failed to receive message",
+                    )
+                    .into());
+                }
+            }
+        }
+
+        let (primary_header, fields_len) = PrimaryHeader::read(&bytes)?;
+        let header_len = MIN_MESSAGE_SIZE + fields_len as usize;
+        let body_padding = padding_for_8_bytes(header_len);
+        let body_len = primary_header.body_len() as usize;
+        let total_len = header_len + body_padding + body_len;
+        if total_len > MAX_MESSAGE_SIZE {
+            return Err(crate::Error::ExcessData);
+        }
+
+        // By this point we have a full primary header, so we know the exact length of the complete
+        // message.
+        bytes.resize(total_len, 0);
+
+        // Now we have an incomplete message; read the rest
+        while pos < total_len {
+            let res = self.recvmsg(&mut bytes[pos..]).await?;
+            let read = {
+                #[cfg(unix)]
+                {
+                    fds.extend(res.1);
+                    res.0
+                }
+                #[cfg(not(unix))]
+                {
+                    res
+                }
+            };
+            pos += read;
+            if read == 0 {
+                return Err(crate::Error::InputOutput(
+                    std::io::Error::new(
+                        std::io::ErrorKind::UnexpectedEof,
+                        "failed to receive message",
+                    )
+                    .into(),
+                ));
+            }
+        }
+
+        // If we reach here, the message is complete; return it
+        let endian = Endian::from(primary_header.endian_sig());
+        let ctxt = Context::new_dbus(endian, 0);
+        #[cfg(unix)]
+        let bytes = serialized::Data::new_fds(bytes, ctxt, fds);
+        #[cfg(not(unix))]
+        let bytes = serialized::Data::new(bytes, ctxt);
+        Message::from_raw_parts(bytes, seq)
+    }
+
+    /// Attempt to receive bytes from the socket.
     ///
     /// On success, returns the number of bytes read as well as a `Vec` containing
     /// any associated file descriptors.
-    async fn recvmsg(&mut self, buf: &mut [u8]) -> RecvmsgResult;
+    ///
+    /// The default implementation simply panics. Implementers must override either `read_message`
+    /// or this method.
+    async fn recvmsg(&mut self, _buf: &mut [u8]) -> RecvmsgResult {
+        unimplemented!("`ReadHalf` implementers must either override `read_message` or `recvmsg`");
+    }
 
     /// Supports passing file descriptors.
     ///
@@ -159,6 +283,14 @@ pub trait WriteHalf: std::fmt::Debug + Send + Sync + 'static {
 impl ReadHalf for Box<dyn ReadHalf> {
     fn can_pass_unix_fd(&self) -> bool {
         (**self).can_pass_unix_fd()
+    }
+
+    async fn receive_message(
+        &mut self,
+        seq: u64,
+        already_received_bytes: Option<Vec<u8>>,
+    ) -> crate::Result<Message> {
+        (**self).receive_message(seq, already_received_bytes).await
     }
 
     async fn recvmsg(&mut self, buf: &mut [u8]) -> RecvmsgResult {
